requirements.txt
autogen-agentchat
autogen-core
autogen-ext[openai]
opentelemetry-sdk
braintrust
python-dotenv
